CORE-15575: added new foreign exchange sample
