+++ kotlin-samples/foreign-exchange/.ci/nightly/JenkinsfileSnykScan
