+++ kotlin-samples/foreign-exchange/config/static-network-config.json
