+++ kotlin-samples/foreign-exchange/gradle.properties
