+++ kotlin-samples/foreign-exchange/settings.gradle
